requirements.txt
fastapi==0.115.4
python-dotenv==1.0.1
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0
SQLAlchemy==2.0.36
pytest==8.3.3
pytest-asyncio==0.24.0
uvicorn==0.32.0
python-multipart==0.0.17
SQLAlchemy-ImageAttach==1.1.0